+++ packages/datasource-customizer/src/decorators/actions/types/fields.ts
@@ -48,11 +48,11 @@ type FileDynamicField<Context> = BaseDynamicField<'File', Context, File>;
 type FileListDynamicField<Context> = BaseDynamicField<'FileList', Context, File[]>;
 type JsonDynamicField<Context> = BaseDynamicField<'Json', Context, Json>;
 type NumberDynamicField<Context> = BaseDynamicField<'Number', Context, number>;
-
 type NumberListDynamicField<Context> = BaseDynamicField<'NumberList', Context, number[]>;
+type DateDynamicField<Context> = BaseDynamicField<'Date' | 'Dateonly', Context, Date>;
 
 type StringDynamicField<Context> = BaseDynamicField<
-  'Date' | 'Dateonly' | 'String',
+  'String' | 'Date' | 'Dateonly',
   Context,
   string
 >;
@@ -86,6 +86,14 @@ type CheckboxDynamicFieldConfiguration = {
 type TextInputFieldConfiguration = {
   widget: 'TextInput';
   placeholder?: string;
+};
+
+type DatePickerInputFieldConfiguration<Context> = {
+  widget: 'DatePicker';
+  format?: ValueOrHandler<Context, string>;
+  placeholder?: string;
+  min?: ValueOrHandler<Context, Date>;
+  max?: ValueOrHandler<Context, Date>;
 };
 
 type ArrayTextInputFieldConfiguration = {
@@ -173,6 +181,7 @@ export type DynamicField<Context = unknown> = StrictUnion<
   | (NumberListDynamicField<Context> & CheckboxesFieldConfiguration<Context, number>)
   | (NumberListDynamicField<Context> & ArrayNumberInputFieldConfiguration<Context>)
   | StringDynamicField<Context>
+  | (DateDynamicField<Context> & DatePickerInputFieldConfiguration<Context>)
   | (StringDynamicField<Context> & TextInputFieldConfiguration)
   | (StringDynamicField<Context> & DropdownDynamicFieldConfiguration<Context, string>)
   | (StringDynamicField<Context> & DropdownDynamicSearchFieldConfiguration<Context, string>)