+++ packages/datasource-toolkit/CHANGELOG.md
@@ -1,3 +1,10 @@
+# @forestadmin/datasource-toolkit [1.22.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-toolkit@1.21.0...@forestadmin/datasource-toolkit@1.22.0) (2023-09-15)
+
+
+### Features
+
+* support date picker in action form widgets ([#819](https://github.com/ForestAdmin/agent-nodejs/issues/819)) ([f7ead87](https://github.com/ForestAdmin/agent-nodejs/commit/f7ead87f6c4bc3970f07ec7c965725a4c40b4ad6))
+
 # @forestadmin/datasource-toolkit [1.21.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-toolkit@1.20.1...@forestadmin/datasource-toolkit@1.21.0) (2023-09-12)